pyproject.toml
[project]
name = "ai-radio-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "feedparser>=6.0.11",
    "langchain>=0.3.21",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
